arxius correcte 1.0 faltanomes itoa

+++ ft_bzero.c
@@ -6,11 +6,12 @@
 /*   By: msoler-e <msoler-e@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/12 11:41:25 by msoler-e          #+#    #+#             */
-/*   Updated: 2022/01/12 14:27:06 by msoler-e         ###   ########.fr       */
+/*   Updated: 2022/01/21 12:44:01 by msoler-e         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+#include "libft.h"
 
-void	ft_bzero(void *str, unsigned int n)
+void	ft_bzero(void *str, size_t n)
 {
 	unsigned char	*ptr;
